Fix RESTflow global library update
This is a difficult one. I can create new global libraries, but updating
them doesn't work anymore (it does, but IDEA reverts it back shortly after).
This is now a workaround: it simply creates an initial library with JAR
directories, then it keeps those in-sync.

Also, lots of smaller updates, version bumps etc.